[EP-1456] Renamed some component.tsx to ActualName.tsx
