+++ packages/earth-map/src/components/filter-by/index.tsx
@@ -17,4 +17,4 @@
   specific language governing permissions and limitations under the License.
 */
 
-export { default } from './component';
+export { default } from './FilterBy';
